refactor: code를 long 타입으로 변경 (#37)

+++ src/main/java/com/prography/pingpong/dto/response/ApiResponse.java
@@ -4,7 +4,7 @@ import com.prography.pingpong.view.ResponseMessage;
 import org.springframework.http.HttpStatus;
 
 public record ApiResponse(
-        Integer code,
+        long code,
         String message
 ) {
